# backend/collector-openapi.yml
openapi: 3.0.0
components:
  schemas:
...
    ItemsResponse:
      title: ItemsResponse
      description: A array with all items of a collection.
      type: object
      required:
        - collections
        - nextKey
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              title:
                description: >-
                  The title of the item (e.g., the name of the book, movie, or
                  comic).
                type: string
                minLength: 1
              description:
                description: A brief description or summary of the item.
                type: string
                minLength: 1
              itemType:
                description: The type of the item (e.g., book, movie, game, comic).
                type: string
                minLength: 1
              userId:
                type: string
              addedAt:
                type: string
                format: date
              lastModified:
                type: string
                format: date
              attachmentUrl:
                description: >-
                  A upload url to attach a file to an item or location depending
                  on the request parameter.
                type: string
                minLength: 1
              locationId:
                description: A physical location where the item is stored.
                type: string
                minLength: 1
              isLendable:
                description: >-
                  Information about whether the user (owner) would lend the
                  item.
                type: boolean
              ownershipStatus:
                description: >-
                  Information about whether the user owns the item, wishes to
                  buy it, has borrowed it, etc.
                type: string
                enum:
                  - OWNER
                  - BORROWER
                  - WISHLIST
              status:
                description: >-
                  Information about whether the item has been read/watched, is
                  currently being read/watched, or is on a to-read/to-watch
                  list.
                type: string
                enum:
                  - TODO
                  - PROCESSING
                  - DONE
        nextKey:
          type: string
          nullable: true
      additionalProperties: false
    ItemResponse:
      title: ItemResponse
      description: >-
        Stores information about each item that can be part of a user's
        collection.
      required:
        - title
        - isLendable
        - ownershipStatus
        - status
      type: object
      properties:
        itemId:
          type: string
        title:
          description: The title of the item (e.g., the name of the book, movie, or comic).
          type: string
          minLength: 1
        description:
          description: A brief description or summary of the item.
          type: string
          minLength: 1
        itemType:
          description: The type of the item (e.g., book, movie, game, comic).
          type: string
          minLength: 1
        userId:
          type: string
        addedAt:
          type: string
          format: date
        lastModified:
          type: string
          format: date
        attachmentUrl:
          description: >-
            A upload url to attach a file to an item or location depending on
            the request parameter.
          type: string
          minLength: 1
        locationId:
          description: A physical location where the item is stored.
          type: string
          minLength: 1
        isLendable:
          description: Information about whether the user (owner) would lend the item.
          type: boolean
        ownershipStatus:
          description: >-
            Information about whether the user owns the item, wishes to buy it,
            has borrowed it, etc.
          type: string
          enum:
            - OWNER
            - BORROWER
            - WISHLIST
        status:
          description: >-
            Information about whether the item has been read/watched, is
            currently being read/watched, or is on a to-read/to-watch list.
          type: string
          enum:
            - TODO
            - PROCESSING
            - DONE
      additionalProperties: false
...
    ErrorResponse:
      title: ErrorResponse
      description: Represents an error.
      required:
        - statusCode
        - errorCode
        - message
      type: object
      properties:
        statusCode:
          description: The http status code.
          type: number
          minimum: 100
          maximum: 599
        errorCode:
          description: An unique error code.
          type: string
          minLength: 1
        message:
          description: Short error message.
          type: string
          minLength: 1
        details:
          description: Additional details about the error (optional).
          type: string
          minLength: 1
      additionalProperties: false
    limit:
      type: string
    nextKey:
      type: string
    Access-Control-Allow-Origin:
      type: string
      default: '*'
      example: https://developer.mozilla.org
    Access-Control-Allow-Credentials:
      type: boolean
      default: true
...
    48aedce9-58d7-4570-a510-cb4598bf5f8b:
      title: CreateItemRequest
      description: >-
        Stores information about each item that can be part of a user's
        collection.
      required:
        - title
        - isLendable
        - ownershipStatus
        - status
      type: object
      properties:
        title:
          description: The title of the item (e.g., the name of the book, movie, or comic).
          type: string
          minLength: 1
        itemType:
          description: The type of the item (e.g., book, movie, game, comic).
          type: string
          minLength: 1
        description:
          description: A brief description or summary of the item.
          type: string
          minLength: 1
        locationId:
          description: A physical location where the item is stored.
          type: string
          minLength: 1
        isLendable:
          description: Information about whether the user (owner) would lend the item.
          type: boolean
        ownershipStatus:
          description: >-
            Information about whether the user owns the item, wishes to buy it,
            has borrowed it, etc.
          type: string
          enum:
            - OWNER
            - BORROWER
            - WISHLIST
        status:
          description: >-
            Information about whether the item has been read/watched, is
            currently being read/watched, or is on a to-read/to-watch list.
          type: string
          enum:
            - TODO
            - PROCESSING
            - DONE
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
info:
  title: Collector API
  description: Serverless application for collection management
  version: v.0.0.0
paths:
...
  /items:
    get:
      summary: Get all items
      description: Get all itemss of a user
      operationId: GetItems
      parameters:
        - name: limit
          in: query
          description: (optional) Iterates over the items with the given limit.
          required: false
          schema:
            $ref: '#/components/schemas/limit'
        - name: nextKey
          in: query
          description: (optional) Iterates over the items with the given nextKey.
          required: false
          schema:
            $ref: '#/components/schemas/nextKey'
      tags:
        - item
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A array with all items of a user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemsResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '400':
          description: Bad Request - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '401':
          description: Unauthorized - user is not logged in or authorization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '500':
          description: Internal Server Error - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
    post:
      summary: Create a new item
      description: Creates a new item for a user
      operationId: CreateItem
      parameters: []
      tags:
        - item
      security:
        - bearerAuth: []
      requestBody:
        description: ''
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/48aedce9-58d7-4570-a510-cb4598bf5f8b'
      responses:
        '201':
          description: The newly created item with additional parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '400':
          description: Bad Request - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '401':
          description: Unauthorized - user is not logged in or authorization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '500':
          description: Internal Server Error - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
...